cli: add `jj op integrate` command for adding operation to op log
When `.jj/working_copy/checkout` refers to an operation that's not in
the op heads, there can be a few different cases. It's typically
simply behind or ahead, but it can also be a sibling. Until we
implement `--no-commit-transaction`, it would indicate a bug if that
happens. However, it does happen sometimes (rarely) at Google where we
can fail to flush the update of the op heads to the server for some
reason, but since we optimistically assume that it will succeed, we
update the working copy to refer to the operation before we have
flushed the op-head update. When in this state, we currently print an
error saying that the repo operation is a sibling of the working
copy's operation, but we don't provide a way of fixing it. This patch
provides a command for fixing this state. It will also be useful once
we have `--no-commit-transaction` (which will presumably print the
un-integrated operation id).

We currently call the process of adding an operation to the operation
heads "publish" in the code. I still went with "integrate" here
because I think it's clearer. If we go with "integrate", we can rename
the internal uses of "publish" to match in a follow-up. Other
reasonable verbs include "record" and "accept". Or we can of course go
with "publish".