eol: fix binary check on files with a CRLF at the probe boundary
eol.rs checks the first 8K bytes of a file to see if it's binary or not.
If there is a null byte found, or a carriage return not followed by a newline, it's considered binary.

This breaks in cases where the 8K boundary is the middle of a CRLF pair, as the carriage return isn't followed by a newline (in the truncated data).