metaedit: fix timestamp in help text
rfc2822, and by extension chrono's parser for rfc2822, only supports a
limited number of timezones (4.3, "Obsolete Date and Time"). JST is not
one of the enumerated zones.

Current fix is to just use a supported time zone in the example. It's
still possible to use a non-enumerated timezone with a human-readable
format by giving the timezone offset directly (`-0800` for PST).

fixes #8421